Fix structure and delete duplicate
