--- reverse/client/MissyCore/_model/_model/__init__.py
@@ -1,7 +1,0 @@
-__import__('pkg_resources').declare_namespace(__name__)
-
-from .assignation import Assignation
-from .event import Event
-from .missyCore import MissyCore
-from .target import Target
-from .user import User
