--- reverse/client/MissyCore/_model/_model/assignation.py
@@ -1,19 +1,0 @@
-from .target import Target
-
-class Assignation():
-
-    def __init__(self, id: int):
-        self.id = id
-
-    def splitAssignations(self, assignation):
-        pass
-        """ for _assignation in assignation:
-
-            if(_assignation[0] == "Animateur"):
-                self.animateur = _assignation[1]
-            elif(_assignation[0] == "Secretaire"):
-                self.secretaire = _assignation[1]
-            elif(_assignation[0] == "Scribe"):
-                self.scribe = _assignation[1]
-            elif(_assignation[0] == "Gestionnaire"):
-                self.gestionnaire = _assignation[1] """
