--- reverse/client/MissyCore/_model/_model/user.py
@@ -1,4 +1,0 @@
-class User(): 
-
-	def __init__(self) -> None:
-		pass
